# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MovieFlix API",
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Listar filmes",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "tags": [
                    "Movie"
                ],
                "summary": "Atualizar filme",
                "description": "Atualiza as informações de um filme existente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                
                "responses": {
                    "204": {
                        "description": "Filme atualizado com sucesso"
                    },
                    "404": {
                        "description": "Filme não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movie"
                ],
                "summary": "Deletar filme",
                "description": "Deleta um filme existente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme. **Atenção:** Este campo exige um formato UUID válido. Requisições com formatos incorretos serão rejeitadas pelo servidor com erro 400 (Bad Request).",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filme deletado com sucesso"
                    },
                    "404": {
                        "description": "Filme não encontrado"
                    },
                    "500": {
                        "description": "Não foi possível deletar o filme"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "example": "Inception"
                    },
                    "genre": {
                        "type": "string",
                        "example": "Ação"
                    },
                    "language": {
                        "type": "string",
                        "example": "Inglês"
                    },
                    "oscar_count": {
                        "type": "integer"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": [
                    "title",
                    "genre_id",
                    "language_id",
                    "oscar_count",
                    "release_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Inception"
                    },
                    "genre_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "Ação"
                    },
                    "language_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "Inglês"
                    },
                    "oscar_count": {
                        "type": "integer"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date"
                    }
                }
            }
        }
    }
}